Updated delete for schedule and relations M:M Applied

+++ src/controllers/scheduleControl.js
@@ -5,6 +5,7 @@ const cors = require('cors');
 const { Op } = require("sequelize");
 
 const Schedule=require('../models/Schedule');
+const ProSchedule=require('../models/ManyToMany_Models/ProductSchedule');
 const AdjustTime=require('../middleware/AdjustTime');
 
 router.use(cors())
@@ -17,8 +18,14 @@ router.post('/register',AdjustTime,async(req,res)=>{
 			ClientName:req.body.ClientName,
 			userId:req.body.userId
 		})
-		if(result){
-			res.json({menssage:"Sucess on Create"})
+		if(result){	
+			const resp=await ProSchedule.create({
+				ScheduleId:result.id,
+				ProSerId:req.body.ProSerId
+			})
+			if(resp){
+				res.json({menssage:"Sucess on Create"})
+			}
 		}
 	}catch(err){
 		res.status(400).send({Error:"Creation Failed"});
@@ -65,16 +72,26 @@ router.get('/getAllFromDay',async(req,res)=>{
 
 router.delete('/delete/One',async(req,res)=>{
 	try{
-		const del=await Schedule.destroy({
+		const find=await Schedule.findOne({
 			where:{
 				[Op.and]:[{ScheduledDay:req.body.ScheduledDay },{ScheduledHour:req.body.ScheduledHour },{userId:req.body.userId}]
 			}
 		})
-		if(del){
-			res.json({menssage:"Shedule deleted"})
-		}
-		else{
-			res.json({Error:"Schedule not deleted"})
+		if(find){
+			const DelX=await ProSchedule.destroy({
+				where:{ScheduleId:find.id}
+			})
+			if(DelX){
+				const del=await Schedule.destroy({
+					where:{id:find.id}
+				})
+				if(del){
+					res.json({menssage:"Shedule deleted"})
+				}
+				else{
+					res.json({Error:"Schedule not deleted"})
+				}
+			}
 		}
 	}catch(err){
 		console.log(err)
@@ -82,25 +99,41 @@ router.delete('/delete/One',async(req,res)=>{
 	}
 })
 
-router.delete('/delete/Day',async(req,res)=>{
+router.delete('/delete/EntireDay',async(req,res)=>{
 	try{
-		const del=await Schedule.destroy({
+		const find=await Schedule.findAll({
 			where:{
 				[Op.and]:[{ScheduledDay:req.body.ScheduledDay }, { userId:req.body.userId }]
-			}
+			},raw:true
 		});
-		if(del){
-			res.json({menssage:"Records Deleted Successfully"});
-		}
-		else{
-			res.json({menssage:"Records not Deleted"});
+
+		if(find){
+			const data=await find.map(function(item,ID){
+				return item.id
+			})
+			for(var x=0;x<data.length;x++){
+				const delx=await ProSchedule.destroy({
+					where:{ScheduleId:data[x]}
+				})
+			}
+			const del=await Schedule.destroy({
+				where:{
+					[Op.and]:[{ScheduledDay:req.body.ScheduledDay }, { userId:req.body.userId }]
+				}
+			});
+			if(del){
+				res.json({menssage:"Records Deleted Successfully"});
+			}
+			else{
+				res.json({menssage:"Records not Deleted"});
+			}
 		}
 	}catch(err){
 		res.status(400).send({Error:"Error"});
 	}
 })
 
-router.put('/update/Day',AdjustTime,async(req,res)=>{
+router.put('/update',AdjustTime,async(req,res)=>{
 	try{
 		const result=await Schedule.findOne({
 			where:{
